Updated the recording and the replay services.

+++ src/Application/LeagueRecorder.Windows/Recording/SpectatorService.cs
@@ -1,22 +1,46 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
+using System.Net.Http;
 using System.Threading.Tasks;
+using LeagueRecorder.Abstractions.Data;
 using LeagueRecorder.Abstractions.Recording;
 using LeagueRecorder.Windows.Extensions;
+using LiteGuard;
+using NeverNull;
 
 namespace LeagueRecorder.Windows.Recording
 {
     public class SpectatorService : ISpectatorService
     {
-        public async Task ExecuteSpectatorCommandsAsync(string commands)
+        public async Task<bool> SpectateMatchAsync(MatchInfo match)
         {
+            Guard.AgainstNullArgument("match", match);
+
+            HttpResponseMessage response = await this.CreateClient(match.Region)
+                                                     .GetAsync(string.Format("/match/observer/id={0}", match.GameId))
+                                                     .ConfigureAwait(false);
+
+            if (response.IsSuccessStatusCode == false)
+                return false;
+
+            string commands = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+
             string filePath = await CreateBatchFile(commands).ConfigureAwait(false);
 
             await Process.Start(filePath).WaitForExitAsync().ConfigureAwait(false);
             File.Delete(filePath);
+
+            return true;
         }
 
+        private HttpClient CreateClient(Region region)
+        {
+            return new HttpClient
+            {
+                BaseAddress = new Uri(region.GetBaseUri())
+            };
+        }
         private static async Task<string> CreateBatchFile(string commands)
         {
             string fileName = string.Format("{0}.bat", Guid.NewGuid().ToString("N"));
